+++ NotoriousTest.Database/Settings/DatabaseSettings.cs
@@ -1,0 +1,7 @@
+﻿namespace NotoriousTest.Database.Settings
+{
+    public class DatabaseSettings
+    {
+        public string ConnectionString { get; set; }
+    }
+}