+++ NotoriousTest.PostgreSql/PostgreInfrastructure.cs
@@ -1,17 +1,16 @@
-﻿using Microsoft.Data.SqlClient;
+﻿using NotoriousTest.Database;
+using NotoriousTest.Database.Settings;
 
-using NotoriousTest.Database;
+using Npgsql;
 
 using Respawn;
 using Respawn.Graph;
 
 using System.Data.Common;
 
-using Testcontainers.MsSql;
-
-namespace NotoriousTest.SqlServer
+namespace NotoriousTest.PostgreSql
 {
-    public class SqlServerContainerInfrastructure : SqlServerContainerInfrastructure<string>
+    public class PostgreInfrastructure : PostgreInfrastructure<string, DatabaseSettings>
     {
         /// <summary>
         /// Gets the configuration key used to retrieve the default connection string.
@@ -27,45 +26,39 @@ namespace NotoriousTest.SqlServer
         }
     }
 
-    public class SqlServerContainerInfrastructure<TOutputConfiguration> : DockerDatabaseInfrastructure<MsSqlContainer, TOutputConfiguration>
+    public class PostgreInfrastructure<TOutputConfiguration, TSettings> : ExternalDatabaseInfrastructure<TOutputConfiguration, TSettings> where TSettings : DatabaseSettings, new()
     {
         public string[] SchemasToInclude { get; init; } = [];
         public string[] SchemasToExclude { get; init; } = [];
 
-        public SqlServerContainerInfrastructure() : base()
+        public PostgreInfrastructure() : base()
         {
-            Container = ConfigureSqlContainer(new MsSqlBuilder()).Build();
             EnsureExtension(new RespawnExtension(() => new RespawnerOptions()
             {
                 TablesToIgnore = TableToIgnore.Select(tti => new Table(tti)).ToArray(),
                 TablesToInclude = TableToInclude.Select(tti => new Table(tti)).ToArray(),
                 SchemasToExclude = SchemasToExclude,
                 SchemasToInclude = SchemasToInclude,
-                DbAdapter = DbAdapter.SqlServer
+                DbAdapter = DbAdapter.Postgres
             }));
-        }
-
-        protected virtual MsSqlBuilder ConfigureSqlContainer(MsSqlBuilder builder)
-        {
-            return builder;
         }
 
         public override DbConnection GetDatabaseConnection()
         {
-            return new SqlConnection(GetDatabaseConnectionString());
+            return new NpgsqlConnection(GetDatabaseConnectionString());
         }
 
         public override DbConnection GetServerConnection()
         {
-            return new SqlConnection(GetServerConnectionString());
+            return new NpgsqlConnection(GetServerConnectionString());
         }
 
         public override string GetDatabaseConnectionString()
         {
-            SqlConnectionStringBuilder connectionString = new SqlConnectionStringBuilder(Container.GetConnectionString());
+            NpgsqlConnectionStringBuilder connectionString = new NpgsqlConnectionStringBuilder(Settings.ConnectionString);
             if (!string.IsNullOrEmpty(FullDbName))
             {
-                connectionString.InitialCatalog = FullDbName;
+                connectionString.Database = FullDbName;
             }
 
             return connectionString.ToString();
@@ -75,10 +68,18 @@ namespace NotoriousTest.SqlServer
         {
             using (DbCommand command = sqlConnection.CreateCommand())
             {
-                command.CommandText = $"CREATE DATABASE [{FullDbName}]";
+                command.CommandText = $"CREATE DATABASE \"{FullDbName}\"";
                 await command.ExecuteNonQueryAsync();
             }
         }
 
+        protected override async Task DropDatabase(DbConnection sqlConnection)
+        {
+            using (DbCommand command = sqlConnection.CreateCommand())
+            {
+                command.CommandText = $"DROP DATABASE \"{FullDbName}\"";
+                await command.ExecuteNonQueryAsync();
+            }
+        }
     }
 }