+++ NotoriousTest.UnitTests/MultipleInfrastructureTestCasesEnvironment.cs
@@ -1,23 +1,25 @@
-namespace NotoriousTest.UnitTests
+using NotoriousTest.Infrastructures;
+
+namespace NotoriousTest.UnitTests;
+
+public partial class AsyncEnvironmentUnitTests
 {
-
-    public partial class AsyncEnvironmentUnitTests
+    public class MultipleInfrastructureTestCasesEnvironment : Environments.Environment
     {
-        public class MultipleInfrastructureTestCasesEnvironment : Environments.Environment
+        private readonly Infrastructure _infra1;
+        private readonly Infrastructure _infra2;
+
+        public MultipleInfrastructureTestCasesEnvironment(Infrastructure infra1, Infrastructure infra2)
         {
+            _infra1 = infra1;
+            _infra2 = infra2;
+        }
 
-            public MultipleInfrastructureTestCasesEnvironment()
-            {
-
-            }
-
-            public override Task ConfigureEnvironment()
-            {
-                AddInfrastructure(_multipleInfrastructureTestCasesInfrastructure1);
-                AddInfrastructure(_multipleInfrastructureTestCasesInfrastructure2);
-
-                return Task.CompletedTask;
-            }
+        public override Task ConfigureEnvironment()
+        {
+            AddInfrastructure(_infra1);
+            AddInfrastructure(_infra2);
+            return Task.CompletedTask;
         }
     }
 }