+++ NotoriousTest.UnitTests/MultipleInfrastructureTestCasesInfrastructure2.cs
@@ -1,17 +1,6 @@
-using NotoriousTest.Infrastructures;
+namespace NotoriousTest.UnitTests;
 
-namespace NotoriousTest.UnitTests
+public partial class AsyncEnvironmentUnitTests
 {
-
-    public partial class AsyncEnvironmentUnitTests
-    {
-        public abstract class MultipleInfrastructureTestCasesInfrastructure2 : Infrastructure
-        {
-
-            public MultipleInfrastructureTestCasesInfrastructure2() : base(false)
-            {
-
-            }
-        }
-    }
+    public abstract class MultipleInfrastructureTestCasesInfrastructure2 : FakeInfrastructure { }
 }