+++ NotoriousTest.UnitTests/Object/ConfigurationObjectTestCasesInfrastructure1.cs
@@ -1,35 +1,22 @@
 using NotoriousTest.Infrastructures;
 
-namespace NotoriousTest.UnitTests
+namespace NotoriousTest.UnitTests;
+
+public partial class AsyncEnvironmentUnitTests
 {
-
-    public partial class AsyncEnvironmentUnitTests
+    public class ConfigurationObjectTestCasesInfrastructure1 : Infrastructure<ConfigurationObjectTestCasesInfrastructureConfiguration1>
     {
-        public class ConfigurationObjectTestCasesInfrastructure1 : Infrastructure<ConfigurationObjectTestCasesInfrastructureConfiguraton1>
+        public override Task Destroy() => Task.CompletedTask;
+
+        public override Task Initialize()
         {
-            public ConfigurationObjectTestCasesInfrastructure1() : base(false)
+            AddEntry(nameof(ConfigurationObjectTestCasesInfrastructureConfiguration1), new()
             {
-            }
-
-
-            public override Task Destroy()
-            {
-                return Task.CompletedTask;
-            }
-
-            public override Task Initialize()
-            {
-                AddOutputConfigurationEntry(new()
-                {
-                    Key1 = "Infra1Key1"
-                });
-                return Task.CompletedTask;
-            }
-
-            public override Task Reset()
-            {
-                return Task.CompletedTask;
-            }
+                Key1 = "Infra1Key1"
+            });
+            return Task.CompletedTask;
         }
+
+        public override Task Reset() => Task.CompletedTask;
     }
 }